requirements.txt
fastapi==0.124.4
uvicorn[standard]
sqlmodel
jinja2
psutil>=5.9.0
paho-mqtt>=2.0.0
httpx
alembic
python-multipart
pyyaml
websockets
wsproto
bcrypt
python-dotenv
pytest
pytest-cov
paramiko
requests
aiohttp
